fix: Table Component
