+++ src/components/UserManagement/ShowTableComponent/styles.js
@@ -1,0 +1,8 @@
+import { makeStyles } from '@material-ui/core'
+
+export const showTableStyles = makeStyles((theme) => ({
+  margin: {
+    margin: theme.spacing(1),
+    justify: 'center',
+  },
+}))
