+++ src/components/UserManagement/TableComponent/TableHeadComponent/styles.js
@@ -1,0 +1,17 @@
+import { makeStyles } from '@material-ui/core/styles'
+
+export const useTableHeadStyles = makeStyles(()=>({
+  
+  visuallyHidden:{
+    border: 0,
+    clip: 'rect(0 0 0 0)',
+    height: 1,
+    margin: -1,
+    overflow: 'hidden',
+    padding: 0,
+    position: 'absolute',
+    top: 20,
+    width: 1,
+  },
+}))
+
